+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - static management page
 - StaticMember model
 - creating static member for new characters statics
+- rendering statics for user with membership
 
 ## [0.6.0] - 2019-11-01
 ### Added